server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sock==0.7.0
aiohttp==3.9.1
python-dotenv==1.0.0
